requirements.txt
grpcio==1.59.0
grpcio-tools==1.59.0
protobuf==4.25.1
asyncio-mqtt==0.16.1
pytest==7.4.3
pytest-asyncio==0.21.1
